+++ ml_tools/ML_evaluation/_autoencoder_metrics.py
@@ -20,7 +20,7 @@ from sklearn.metrics import (
 
 from ..ML_configuration._config_metrics import FormatAutoencoderMetrics
 
-from ..keys._keys import _EvaluationConfig
+from ..keys._config import _EvaluationConfig
 from ..path_manager import make_fullpath, sanitize_filename
 from .._core import get_logger
 from .._helpers import wrap_text