+++ ml_tools/ML_evaluation/_eval_classification.py
@@ -30,7 +30,7 @@ from ..ML_configuration._config_metrics import (_BaseMultiLabelFormat,
 from ..path_manager import make_fullpath, sanitize_filename
 from .._core import get_logger
 from .._helpers import wrap_text
-from ..keys._keys import _EvaluationConfig
+from ..keys._config import _EvaluationConfig
 
 from ._radar_plots import (
     mpl_to_plotly_rgba,