+++ ml_tools/ML_evaluation_captum/_feature_importance_sequence.py
@@ -11,7 +11,9 @@ from captum.attr import IntegratedGradients
 from ..path_manager import make_fullpath, sanitize_filename
 from .._core import get_logger
 from .._helpers import wrap_text
-from ..keys._keys import CaptumKeys, _EvaluationConfig
+from ..keys._keys import CaptumKeys
+from ..keys._config import _EvaluationConfig
+
 
 _LOGGER = get_logger("Captum Sequence")